Cache edits made with PolyGerrit
Currently, it is quite easy to lose edits in process, as they are not
stored anywhere and there is no prompt before navigation away from
editing.

With this change, the editor view will cache edits to gr-storage at a
debounced rate, similar to gr-editable-content and gr-diff-comment.

These edits are keyed with the change number, patch number, and path of
the current file.

Bug: Issue 8481
Change-Id: I5de55e858225b6518ff5ff37f5b87cebdb84325a

Former-commit-id: aa4a07e8934e14dd90c6fa0a7516852f3db25149