Fix index ssh commands with elasticsearch type
Previously, executing the index changes and project ssh commands on the
elasticsearch index type Gerrit site would result in the error: "fatal:
gerrit: index not found". This change modifies the condition that
restricts the indexing of changes and projects to only Lucene type site
so that the index commands are available for the elasticsearch type
site.

Bug: Issue 8690
Change-Id: I139ffad045b4d855f7b341c769fa38c3831cd58c

Former-commit-id: 06b41b422d41217b5586c771d4e71da65c3d8919