Remove CURRENT_COMMIT from default SubmittedTogether opts
SubmittedTogether potentially returns hundreds of changes.
CURRENT_COMMIT requires opening the repository that contains the change
to load more information about the commit. This can make the endpoint
very slow (>1 minute for 500 changes).

SubmittedTogether already provides a way of specifying additional
options for ChangeJson. Therefore removing an option from the defaults
seems safe as users that still want to receive commit information can
just specify o=CURRENT_COMMIT on the request.

This commit specifies this option for the GWT UI so that we don't run
into (potential) issues with WebLinks. Since the GWT UI will be deleted
soon, keeping it on the slow version of the endpoint seems OK.

This commit also implements a lazyload boolean in SubmittedTogether.
This forces the default call to not do any lazyloading which helps us to
catch issues earlier. The boolean is updated if the user specified an
option that requires lazy loading.

Change-Id: I763d15447fa79846fe0656a869b15edcf8502cb5

Former-commit-id: 6fbbdaa4f677c3ffceb61a087b269a3ce37b873b