Send newchange email to reviewers when review is started
Previously, when a user would add a reviewer to their WIP change while
simultaneously starting review, the "newchange" email to the new
reviewer would fail to send because the logic controlling the default
notification level of that notification only took into account the
change's former WIP state.

This had two significant consequences. First, the opening notification
for many reviewers using this workflow was a terse email that only
mentioned "This change is ready for review." I've generally merged bug
reports about this into issue 6854. This change does not fix issue 6854,
which is about combining these two emails (which is the opposite of
what this change does). However, it does address the confusion driving
many of those merged issues.

The other significant consequence is mail threading, particularly for
GMail users. The "newchange" email is designed to start the thread. All
other email notifications are meant to follow up to the first one. The
former behavior omitted any initial "newchange" email, so the
notification for the first reviewer added after review was started would
be presented by GMail at the top of the thread (potentially days after
it really started). This is issue 7023, which is fixed by this change.

Bug: Issue 7023
Change-Id: I6d19b0a47809dc4a83b036fc6934b043a9819581

Former-commit-id: 63cc2e82885a568e0f8bd9c88272eb155a5a16fa