Update gr-fixed-panel
Previously, when keepOnScroll was set to true, and the header was not
yet to the top, there was a miscalculation in the 'top' number, casuing
the header to scroll up higher than its gr-fixed-panel parent.

This occurred because calculations for '_topLast' became inaccurate, and
compounded on one another.

This change fixes the problem by utilizing the position of the
gr-fixed-panel element itself to get a top offset, since it is reatively
positioned, this can be computed.

gr-fixed-panel tests were not being run, and they were already failing.
The tests have been updated to relfect how they should behave.

Change-Id: If0156cc7b009498ee63dddf0cef1b7e8d786b17c

Former-commit-id: 6c16e97d520ea3ca312390994e87b79589471c95