Use a common interface for Gerrit permissions
Gerrit has defined permissions for hosts, projects, refs, changes,
labels and plugins. But those permissions are not built on one common
interface even though their behaviors are almost the same. This makes
the life in the PermissionBackend a little bit hard. For example, some
general logic for a permission couldn't be shared with others.

This commit defines the 'GerritPermission' interface and adapts existing
permission class/interface to leverage on this common interface.

Change-Id: I56d5afbf0ddd7703fdbdfa3911929dbfab7ad065

Former-commit-id: 8ded5cb1583ce2b67bff0ecccd5f85d59aeec915