InitIndex: Allow to configure index.maxLimit for Elasticsearch
When using the Elasticsearch secondary index, `index.maxLimit` should
not exceed the value of `index.max_result_window` configured on the
Elasticsearch server.

By default `index.maxLimit` is not set, which effectively means that the
value used is Integer.MAX_VALUE. When Elasticsearch is used out of the
box the default `index.max_result_window` is 10000, and this results in
query failures:

  "Result window is too large, from + size must be less than or equal
   to: [10000] but was [2147483647]."

When the index type is Elasticsearch, prompt for the max limit during
initialization, defaulting to Elasticsearch's default value of 10000.

Bug: Issue 8527
Change-Id: I983a83b5cdd0eaf5ba875c3310dfd484e53fe064

Former-commit-id: 52c2b4afe0fa9a71aaee5f1050837bb288f58b73