Fix star icon not highlighted in change list on UI
The star icon for the starred changes did not show up as such in the
change list because the 'star' field for the change index was not
handled. This change adds the settings for the 'star' field of the
change so that all the starred changes can be seen highlighted in the
change list.

Bug: Issue 8588
Change-Id: I6682e1ed009e9c3b2518d1260f66cddd91dcaa31

Former-commit-id: 4c28986103ae001e48f9c533c6cbf3bbbac02bd5