Use the same border color throughout the app
With this change, all CSS border declarations that specify some shade of
gray are modified to use the one unified --border-color declaration in
app-theme.html.

Further followup with UX is required to ensure that the shade of gray
chosen is the correct one -- for now, the most-used one is used.

Change-Id: I2134ebf4a459ba040e01856c11389780307d8648

Former-commit-id: c8280988539bfc576d4ed5fe554c1f3b58a15480