Rely on params change to set _lastSearchPage
Gerrit remembers the last page that contains a change-list (either a
dashboard or a page with search results) so that users can return there
with the 'u' key.

This is done by saving the last relative URL as a property at the gr-app
level.

The app relied on the 'location-changed' event to set this property, but
in some cases that event is fired before the app params are set. One
such case is on a fresh reload of a search page.

In that case, the _handleSearchPageChange function has no way of knowing
what view is currently loaded, and prematurely exits.

With this change, Gerrit now observes the app params to set the
property.

Bug: Issue 7368
Change-Id: I62ead340677d9a416a4713ff003fc3ef18d3624f

Former-commit-id: c27a23c06b982be8de5fe27db3c1874b21ea094b