Propagate cause of a project not found in cache
Propagate the original exception when a project failed to be loaded
into cache.

That allows the callers to have a better understanding of
what is going on and adjust the compensation logic to recover from it
when possible.

For instance, the caller may need to filter out the project if the
exception isn't a critical one (e.g. RepositoryNotFoundException)

Change-Id: Iaa4d050c376176b60c6feb7ad5cd3cac45f02b17

Former-commit-id: 5c1ff86a08db9675b59a82797b9438487eeafcae