Try to stop plugin if enabling it failed
Without this, a plugin failing to load might break other unrelated
plugins. This occurs because a Guice environment is created and a
serverManager is created but never started.

This should only be triggered when a bogus plugin is installed, which
is more likely to occur during plugin development.

Also initialize the fields outside of the constructor.

Change-Id: I0bcbe331697e320430711a776357800f1f69c8a3

Former-commit-id: f13ee1e78a21ca4d05579e9abf121ebc0509a1cd