AbstractQueryChangesTest: Add back test for visible: predicate
Change Icfcb34efe removed draft workflow and at the same time removed
the 'explicitVisibleTo' test, which was testing visibility of draft
changes. As a result, the only test coverage of the visibleto: query
was removed.

Add 'explicitVisibleTo' back, but rename it to 'visible'. The implicit
visibility is already covered in the 'byPrivate' test.

Change-Id: I9be04156ce4401754cebb0c969d1b233a132735e

Former-commit-id: a540ac2c87f8fa2cb1c5bcc37b82ae85656461f0