Set query result title in async
Like the dashboard view, the _paramsChanged method may be called before
the element has been attached. When this happens, the title-change event
is fired, but not intercepted by any parents. Fire the event in an async
in the same way as the dashboard view so that attachment is allowed to
happen first.

Bug: Issue 8769
Change-Id: Id5417633c3f85f23fabb0bc3f6175400bd81803d

Former-commit-id: ac9fd50c0a42033c5b3bd35c1c969529baeea2f0