Fix flaky Elasticsearch tests
The tests were failing sometimes with a failure to create an index
which already existed. Instead of creating and deleting indices between
every test, create unique indices by prefixing the indices name with
test method name. The indices are now clean up after all the tests are
executed, when the server is stopped and temp folder deleted.

Change-Id: I6539845e09af56183dd315f112a35d572db7c920

Former-commit-id: 48e7d8b9a8daf247a7c7a7b2230fa581cd889cff