Extend test API for groups and add tests for it
With this change, the test API for groups allows to update all
properties of internal groups which are modifiable at the moment. In
addition, all properties which may be specified upon group creation are
supported.

To ensure that the test API for groups is working as expected, tests for
it are added. This means that other tests which use this API do not need
to verify themselves that the desired state of the system was obtained.

Change-Id: Iae7547e0f972d8d1eb0c96f8cc25dc701c0c633f

Former-commit-id: ee1e8e57315002642d33617ac036db2ba684f40c