Split setter/getter interfaces out of CacheBinding
The CacheBinding interface now contains only setters. With "binding" in
the name, it indicates that this is the thing used from within
CacheModule to configure the interface.

Move all the getter methods into a new interface, CacheDef. This is the
type that is passed into {Memory,Persistent}CacheFactory, where the
cache setup internals need to query the configured values, but calling
setters at this point would be programmer error.

This has the side advantage of eliminating the need for throwing
RuntimeExceptions from H2CacheBindingProxy, and it is now clearer that
this is simply a delegate for the getter half of the interface.

Both of these interfaces are still implemented by the same CacheProvider
class, and it still retains the freeze-during-get behavior. It may be
possible to split it up, for example converting it to a normal builder
which produces a read-only CacheDef from its build method, but I wasn't
able to get that working with the asEagerSingleton binding magic used by
the current implementation.

Change-Id: I5c90f6eb17d257a127b2c52d70558c2f80655ec2

Former-commit-id: 12b62b053887c28cfd206b09ffeadb74619d9e33