Ensure SubmitRecords are loaded in SubmittedTogether
LocalMergeSuperSet uses InternalChangeQuery to get
byCommitsOnBranchNotMerged. When the number of search
terms exceeds the limit, changes are loaded from the database.

This means, that submitRecords and currentPatchSet are
not populated. Since we set lazyload for ChangeJson,
these fields can be empty when we format the JSON response.

Since maxNumTerms is large, this bug happens very rarely. This
change is a stop-gap until we implement a better solution.

Change-Id: Ica1b01ea45f086b58501040cb6e5bfd0bbaa76b8

Former-commit-id: a39cb382a09993e9597d0eba41c78ba130216e42