Revert migration from refs/for/* permissions to refs/heads/*
The migration contains a bug which causes project.config to become
unreadable, breaking a site until it is fixed. A fix is in progress[1],
but there are still some open questions that we don't have answers to.
Revert the functionality for now, to give us time to fully bake a fix.

This reverts the following commits:
  b85c49ad917b1f9d2be7d9c7850e2b1d3b2901dd [formerly 9ff07177eae4970c1773682cbfc503b329b93e78]
  b69f28962f365eda1f5212b6a3a01982b8e7164e [formerly 8392e982c0fc0f2a73397fab18a381cec050514f]
  a67cfa10bee94eceeec52f3a41525993e6e5f6df [formerly 66c52dae95e16c1bcce9dad4ae91bfc6c0dc04ab]
  ed39d12eef217dc7af3eab0eae4116b2b322c9e9 [formerly 642f3e07e7cb98a997e15569ae01a6952a9bce67]

[1] https://gerrit-review.googlesource.com/c/gerrit/+/173513

Change-Id: I139ba7aa97a946f2c31258e7dfaa3a94ee72ff6e

Former-commit-id: 942a8fe8cfc29a41080c09c04e03ab288b0e487c